IssueNo: https://gitee.com/openharmony/arkui_ace_engine_lite/issues/IAFYD5 describe: There is overflow in the loop variable Feature or Bugfix: Bugfix Binary Source:Yes Signed-off-by: youbing54<youbing3@huawei.com>

+++ frameworks/src/core/modules/presets/cjson_parser.cpp
@@ -426,7 +426,7 @@ bool CJSONParser::IsFileExistFullMatch(const char *fileName)
         return false;
     }
     bool canBeFullMatched = false;
-    for (uint8_t i = 0; i <= count; i++) {
+    for (int i = 0; i <= count; i++) {
         if (!strcmp(fileName, GetNode(fileList, i))) {
             canBeFullMatched = true;
             break;
